misc work

+++ core/include/tangram/platform.h
@@ -108,12 +108,12 @@ public:
     virtual std::vector<FontSourceHandle> systemFontFallbacksHandle() const;
 
     size_t activeUrlRequests() const { return m_urlCallbacks.size(); }
-    void notifyStorage(int dtot, int doffl) const { if(onNotifyStorage) onNotifyStorage(dtot, doffl); }
+    void notifyStorage(int64_t dtot, int64_t doffl) const { if(onNotifyStorage) onNotifyStorage(dtot, doffl); }
 
     std::atomic_bool isOffline{false};
     size_t urlRequestsThreshold = 0;
     std::function<void(void)> onUrlRequestsThreshold;
-    std::function<void(int, int)> onNotifyStorage;
+    std::function<void(int64_t, int64_t)> onNotifyStorage;
 
     static char const* shutdown_message;
     static char const* cancel_message;
